backend/scripts/requirements.txt
pytesseract
Pillow
pymupdf
python-docx